requirements.txt
# Deep Learning et Computer Vision
tensorflow>=2.15.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python
Pillow
scikit-learn>=1.3.0
imbalanced-learn>=0.12.4
matplotlib
seaborn
numpy

